+++ src/ncc/Classes/GithubExtension/GithubService.php
@@ -27,8 +27,8 @@
     use ncc\Classes\HttpClient;
     use ncc\Exceptions\AuthenticationException;
     use ncc\Exceptions\GitException;
-    use ncc\Exceptions\HttpException;
     use ncc\Exceptions\MalformedJsonException;
+    use ncc\Exceptions\NetworkException;
     use ncc\Exceptions\VersionNotFoundException;
     use ncc\Interfaces\RepositorySourceInterface;
     use ncc\Objects\DefinedRemoteSource;
@@ -50,8 +50,8 @@
          * @return RepositoryQueryResults
          * @throws AuthenticationException
          * @throws GitException
-         * @throws HttpException
          * @throws MalformedJsonException
+         * @throws NetworkException
          */
         public static function getGitRepository(RemotePackageInput $packageInput, DefinedRemoteSource $definedRemoteSource, ?Entry $entry = null): RepositoryQueryResults
         {
@@ -82,8 +82,8 @@
          * @return RepositoryQueryResults
          * @throws AuthenticationException
          * @throws GitException
-         * @throws HttpException
          * @throws MalformedJsonException
+         * @throws NetworkException
          * @throws VersionNotFoundException
          */
         public static function getRelease(RemotePackageInput $package_input, DefinedRemoteSource $defined_remote_source, ?Entry $entry = null): RepositoryQueryResults
@@ -98,8 +98,8 @@
          * @return RepositoryQueryResults
          * @throws AuthenticationException
          * @throws GitException
-         * @throws HttpException
          * @throws MalformedJsonException
+         * @throws NetworkException
          * @throws VersionNotFoundException
          */
         public static function getNccPackage(RemotePackageInput $packageInput, DefinedRemoteSource $definedRemoteSource, ?Entry $entry = null): RepositoryQueryResults
@@ -116,8 +116,8 @@
          * @return array
          * @throws AuthenticationException
          * @throws GitException
-         * @throws HttpException
          * @throws MalformedJsonException
+         * @throws NetworkException
          */
         private static function getReleases(RemotePackageInput $packageInput, DefinedRemoteSource $definedRemoteSource, ?Entry $entry = null): array
         {
@@ -188,8 +188,8 @@
          * @return array
          * @throws AuthenticationException
          * @throws GitException
-         * @throws HttpException
          * @throws MalformedJsonException
+         * @throws NetworkException
          */
         private static function getJsonResponse(HttpRequest $httpRequest, ?Entry $entry): array
         {
@@ -215,8 +215,8 @@
          * @return mixed
          * @throws AuthenticationException
          * @throws GitException
-         * @throws HttpException
          * @throws MalformedJsonException
+         * @throws NetworkException
          * @throws VersionNotFoundException
          */
         private static function processReleases(RemotePackageInput $packageInput, DefinedRemoteSource $definedRemoteSource, ?Entry $entry): mixed