+++ src/ncc/Classes/GitlabExtension/GitlabService.php
@@ -26,8 +26,8 @@
     use ncc\Classes\HttpClient;
     use ncc\Exceptions\AuthenticationException;
     use ncc\Exceptions\GitException;
-    use ncc\Exceptions\HttpException;
     use ncc\Exceptions\MalformedJsonException;
+    use ncc\Exceptions\NetworkException;
     use ncc\Exceptions\NotSupportedException;
     use ncc\Exceptions\ResourceNotFoundException;
     use ncc\Interfaces\RepositorySourceInterface;
@@ -51,8 +51,8 @@
          * @return RepositoryQueryResults
          * @throws AuthenticationException
          * @throws GitException
-         * @throws HttpException
          * @throws MalformedJsonException
+         * @throws NetworkException
          */
         public static function getGitRepository(RemotePackageInput $packageInput, DefinedRemoteSource $definedRemoteSource, ?Entry $entry=null): RepositoryQueryResults
         {
@@ -93,8 +93,8 @@
          * @return RepositoryQueryResults
          * @throws AuthenticationException
          * @throws GitException
-         * @throws HttpException
          * @throws MalformedJsonException
+         * @throws NetworkException
          * @throws ResourceNotFoundException
          */
         public static function getRelease(RemotePackageInput $package_input, DefinedRemoteSource $defined_remote_source, ?Entry $entry = null): RepositoryQueryResults
@@ -186,8 +186,8 @@
          * @return array
          * @throws AuthenticationException
          * @throws GitException
-         * @throws HttpException
          * @throws MalformedJsonException
+         * @throws NetworkException
          */
         private static function getReleases(string $owner, string $repository, DefinedRemoteSource $definedRemoteSource, ?Entry $entry): array
         {